requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose
cryptography
pydantic